initial setup of rough structure
After taking about an hour to understand the requirements and to figure out
how exactly the tax calculations worked (math is not my strong suit!), I created
the rough structure of the app. I'd like this project to incorporate some
object-oriented design, with a lot of the tax calculations done on the
TaxCalculator objects.