delete BaseTaxCalculator class
I was originally going to have this as the base tax calculator class
that the other calculators inherit from, but in the interest of time,
and also given that for this coding challenge the tax variations are
not that complicated, I figured it'd be better to just have the
separate tax calculators not inherit from anything. But in the future,
it might be worth having a base master class that other tax
calculators down the road can inherit from.