setup basic structure for the various tax calculators
These are pretty simple tax calculator classes with not much logic in them.
Originally, I was thinking about adding the tax strategy to the item itself
(and we could have multiple different types of items). However, given the
requirements seem to deal more with different combos of tax calculators, I
figured it'd be easier later down the road to extend more upon the tax
calculator logic, vs. trying to configure that in the various items
themselves.