add Item spec
This simple spec file gives me confidence that my setup is working
(and it helped me catch a bug that I documented in the previous commit!).
Normally my 'describe' blocks would focus on just one method, but given
that the total price and total tax methods are so intertwined with
each other, I figured it'd be easier to just test the changes in the
same test block.